+++ src/views/Bond/BondRedeem.tsx
@@ -5,8 +5,8 @@ import { useDispatch, useSelector } from 'react-redux';
 import { BondKey, getBond } from 'src/constants';
 import { prettifySeconds, prettyVestingPeriod, trim } from '../../helpers';
 import { useWeb3Context } from '../../hooks';
-import { redeemBond } from '../../store/slices/bond-slice';
-import { IPendingTxn, isPendingTxn, txnButtonText } from '../../store/slices/pending-txns-slice';
+import { redeemBond } from '../../store/actions/bond-action';
+import { IPendingTxn } from '../../store/slices/pending-txns-slice';
 import { IReduxState } from '../../store/slices/state.interface';
 import { useTranslation, Trans } from 'react-i18next';
 import BondRedeemDialog from './BondRedeemDialog';
